Revised files for v0.1.0 release
(1) Corrected split_sample (util.py) to ensure that values that are repeated several times belong to the same group (this function is used to split to input-output dataset in the PAWN method). (2) Solved the numba incompatibility issues: made the code compatible with numba in hbv_sim (HBV.py, note that a line of the code needs to be changed to run it with Python 2, see notes in the module), sobol_g_function (sobol_g.py) and empiricalcdf (util.py) and removed the use of numba in FAST.py. (3) Made small adjustments to the workflows.